# nodejs-client/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rakam API Documentation",
    "description": "An analytics platform API that lets you create your own analytics services.",
    "version": "1.0.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
  "/recipe/export": {
    
    "post": {
      "summary": "Export recipe",
      "description":"",
      "x-swagger-router-controller": "Recipe",
      "tags": ["Recipe"],
      "operationId": "export",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/Recipe"
  }
}
        
        
      }
    } 
    
  } ,
...
  "/materialized-view/schema": {
    
    "post": {
      "summary": "Get schemas",
      "description":"",
      "x-swagger-router-controller": "Materializedview",
      "tags": ["Materializedview"],
      "operationId": "schema",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "type" : "array",
    "items" : {
      "$ref" : "#/definitions/MaterializedViewSchema"
    }
  }
}
        ,
        "400": {
  "description" : "Project does not exist."
}
        
        
      }
    } 
    
  } ,
  "/materialized-view/delete": {
    
    "post": {
      "summary": "Delete materialized view",
      "description":"",
      "x-swagger-router-controller": "Materializedview",
      "tags": ["Materializedview"],
      "operationId": "delete",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "name",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/JsonResponse"
  }
}
        ,
        "400": {
  "description" : "Project does not exist."
}
        
        
      }
    } 
    
  } ,
  "/materialized-view/create": {
    
    "post": {
      "summary": "Create view",
      "description":"",
      "x-swagger-router-controller": "Materializedview",
      "tags": ["Materializedview"],
      "operationId": "create",
      "parameters": [
        {
  "in" : "body",
  "name" : "MaterializedView",
  "required" : true,
  "schema" : {
    "$ref" : "#/definitions/MaterializedView"
  }
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/JsonResponse"
  }
}
        ,
        "400": {
  "description" : "Project does not exist."
}
        
        
      }
    } 
    
  } ,
  "/materialized-view/list": {
    
    "post": {
      "summary": "List views",
      "description":"",
      "x-swagger-router-controller": "Materializedview",
      "tags": ["Materializedview"],
      "operationId": "listViews",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "type" : "object",
    "properties" : { }
  }
}
        ,
        "400": {
  "description" : "Project does not exist."
}
        
        
      }
    } 
    
  } ,
...
  "/materialized-view/get": {
    
    "post": {
      "summary": "Get view",
      "description":"",
      "x-swagger-router-controller": "Materializedview",
      "tags": ["Materializedview"],
      "operationId": "get",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "name",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "type" : "object",
    "properties" : { }
  }
}
        ,
        "400": {
  "description" : "Project does not exist."
}
        
        
      }
    } 
    
  } ,
...
  }, "definitions": {
  "ContinuousQueryBuilder": {
  "properties" : {
    "name" : {
      "type" : "string"
    },
    "query" : {
      "type" : "string"
    },
    "collections" : {
      "type" : "array",
      "items" : {
        "type" : "string"
      }
    },
    "options" : {
      "type" : "object",
      "additionalProperties" : {
        "type" : "object",
        "properties" : { }
      }
    },
    "tableName" : {
      "type" : "string"
    },
    "partitionKeys" : {
      "type" : "array",
      "items" : {
        "type" : "string"
      }
    }
  }
},"User": {
...
},"Recipe": {
  "required" : [ "collections", "reports", "strategy" ],
  "properties" : {
    "strategy" : {
      "type" : "string",
      "enum" : [ "DEFAULT", "SPECIFIC" ]
    },
    "project" : {
      "type" : "string"
    },
    "collections" : {
      "type" : "object",
      "additionalProperties" : {
        "$ref" : "#/definitions/Collection"
      }
    },
    "reports" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/ReportBuilder"
      }
    },
    "materializedViews" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/MaterializedViewBuilder"
      }
    },
    "continuousQueries" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/ContinuousQueryBuilder"
      }
    },
    "continuousQueryBuilders" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/ContinuousQueryBuilder"
      }
    },
    "materializedViewBuilders" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/MaterializedViewBuilder"
      }
    }
  }
},"BatchCreate": {
...
},"JsonResponse": {
  "properties" : {
    "success" : {
      "type" : "boolean",
      "default" : false
    },
    "message" : {
      "type" : "string"
    }
  }
},"Send": {
...
},"ReportBuilder": {
  "properties" : {
    "slug" : {
      "type" : "string"
    },
    "name" : {
      "type" : "string"
    },
    "query" : {
      "type" : "string"
    },
    "options" : {
      "type" : "object",
      "additionalProperties" : {
        "type" : "object",
        "properties" : { }
      }
    }
  }
},"Measure": {
...
},"MaterializedViewBuilder": {
  "properties" : {
    "name" : {
      "type" : "string"
    },
    "table_name" : {
      "type" : "string"
    },
    "query" : {
      "type" : "string"
    },
    "options" : {
      "type" : "object",
      "additionalProperties" : {
        "type" : "object",
        "properties" : { }
      }
    },
    "updateInterval" : {
      "type" : "string"
    }
  }
},"ABTestingReport": {
...
},"SchemaField": {
  "properties" : {
    "name" : {
      "type" : "string",
      "readOnly" : true
    },
    "type" : {
      "type" : "string",
      "readOnly" : true,
      "enum" : [ "STRING", "LONG", "DOUBLE", "BOOLEAN", "DATE", "TIME", "TIMESTAMP", "ARRAY_STRING", "ARRAY_LONG", "ARRAY_DOUBLE", "ARRAY_BOOLEAN", "ARRAY_DATE", "ARRAY_TIME", "ARRAY_TIMESTAMP" ]
    },
    "nullable" : {
      "type" : "boolean",
      "readOnly" : true,
      "default" : false
    },
    "unique" : {
      "type" : "boolean",
      "readOnly" : true,
      "default" : false
    },
    "descriptiveName" : {
      "type" : "string",
      "readOnly" : true
    },
    "description" : {
      "type" : "string",
      "readOnly" : true
    },
    "category" : {
      "type" : "string",
      "readOnly" : true
    }
  }
},"AutomationRule": {
...
},"MaterializedViewSchema": {
  "properties" : {
    "name" : {
      "type" : "string"
    },
    "fields" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/SchemaField"
      }
    }
  }
},"EventList": {
...
},"Collection": {
  "properties" : {
    "name" : {
      "type" : "string"
    },
    "fields" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/SchemaField"
      }
    }
  }
},"Event": {
...
},"MaterializedView": {
  "properties" : {
    "project" : {
      "type" : "string"
    },
    "name" : {
      "type" : "string"
    },
    "table_name" : {
      "type" : "string"
    },
    "query" : {
      "type" : "string"
    },
    "update_interval" : {
      "type" : "string"
    },
    "options" : {
      "type" : "object",
      "additionalProperties" : {
        "type" : "object",
        "properties" : { }
      }
    }
  }
},"RetentionQuery": {
...
  }
}